// src/app/shared/model/requestDoctorDto.ts
export interface requestDoctorDTO
{
    Id?:string;
    departmentID: string;
    empFName: string;
    empLName: string;
    dateJoining: string;
    empType: string;
    email: string;
    address: string;
    ssn: string;
    scheduleId: string;
    qualifications: string;
    specialization: string;
}